Create 11. Container With Most Water.java
Time Complexity : O(n).
We have given an integer array height of length n. There are n vertical lines drawn such that the two endpoints of the ith line are (i, 0) and (i, height[i]).
We have to find two lines that together with the x-axis form a container, such that the container contains the most water.
For that I have found maximum area by considering x-axis and two bars and solved in O(n) time.