Create 131. Palindrome Partitioning.java
Problem : 
Given a string s, partition s such that every substring of the partition is a palindrome. Return all possible palindrome partitioning of s.
A palindrome string is a string that reads the same backward as forward.

Ex : 
Input: s = "aab"
Output: [["a","a","b"],["aa","b"]]

Time Complexity :- BigO(N*2^N) where N is the length of string s. This is the worst-case time complexity when all the possible substrings are palindrome.
Space Complexity :- BigO(N) as space will be used to store the recursion stack